requirements.txt
certifi==2019.6.16
chardet==3.0.4
dataclasses==0.6
Django==2.2.4
django-jsonfield==1.3.1
django-macaddress==1.6.0
#django-pydash-app==1.4.6
djongo==1.2.33
google-api-python-client==1.4.1
httplib2==0.13.1
idna==2.8
jet-django==0.7.4
netaddr==0.7.19
oauth2client==4.1.3
pyasn1==0.4.6
pyasn1-modules==0.2.6
PyJWT==1.7.1
pymongo==3.9.0
pytz==2019.2
requests==2.22.0
rsa==4.0
six==1.12.0
sqlparse==0.2.4
uritemplate==3.0.0
urllib3==1.25.3
bcrypt
numpy